+++ src/rules.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
  * This code may only be used under the BSD style license found at
  * http://polymer.github.io/LICENSE.txt
  * The complete set of authors may be found at
@@ -12,8 +12,5 @@
  * http://polymer.github.io/PATENTS.txt
  */
 
-import {DomModuleNameOrIs} from '../html/dom-module-name-or-is';
-import {MoveStyleIntoTemplate} from '../html/move-style-into-template';
-
-// Temporary. To replace with a registry.
-export const AllRules = [new DomModuleNameOrIs(), new MoveStyleIntoTemplate()];
+import './html/dom-module-name-or-is';
+import './html/move-style-into-template';