+++ src/polymer-build.ts
@@ -3,6 +3,7 @@ export {SWConfig} from 'sw-precache';
 export {BuildAnalyzer, DepsIndex} from './analyzer';
 export {BuildBundler} from './bundle';
 export {forkStream} from './fork-stream';
+export {HtmlSplitter} from './html-splitter';
 export {PolymerProject} from './polymer-project';
 export {addServiceWorker, generateServiceWorker} from './service-worker';
 export {AddServiceWorkerOptions} from './service-worker';